refactor(vestad): move the Claude OAuth PKCE flow out of docker.rs into providers/claude.rs, its only consumer
